pyproject.toml
[project]
name = "deepagent-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.4.5",
    "dotenv>=0.9.9",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=1.1.10",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
